refactoring the formatters
