+++ modules/formatter/__init__.py
@@ -1,2 +1,3 @@
 from .prompt_builder import *
 from .prompt_parser import *
+from .prompt_formatter import Formatter
