+++ modules/formatter/high_level/formatter_custom.py
@@ -1,4 +1,4 @@
-from .prompt_parser import parse_raw, check_special_tokens
+from modules.formatter.prompt_parser import parse_raw, check_special_tokens
 import copy
